queue: Fix SQL email viewing queue with default columns
This fixes a problem where viewing a queue with no defined columns would
result in attempting to add columns to the queue, which would result in an
incorrect SQL statement and an email sent to the administrator. This was
ultimately moot because columns are not added directly to the queue anymore.